PAGESSSS

+++ test.v1/app/Digitus/View/blog/post.blade.php
@@ -1,73 +1,75 @@
 @extends('layouts.front.blogpost')
 
 @section('title')
-	{{ $post->post_title }}
+	{{ $post->title }}
 @stop
 
 @section('menu')
-	@include('layouts.front.menus.homemenu')
-@stop
-
-@section('sidebar')
-<div id="sidebar" class="sidebar col-sm-2 col-md-2">
-	<ul class="nav nav-sidebar nav-pills nav-stacked">
-		<li><a href="{{{ URL::to('admin/blog/create') }}}" class="btn btn-success btn-xs pull-left"><span class="glyphicon glyphicon-plus"></span> New post</a></li>
-	</ul>
-</div>
+	@include('layouts.front.menus.digimenu')
 @stop
 
 @section('content')
 
-<div class="blogroll container col-sm-offset-2 col-md-offset-2 col-xs-offset-2 col-lg-offset-2">
+<div class="blogroll col-xs-10 col-sm-10 col-md-10 col-lg-10">
 
-		<div class="col-xs-7 col-sm-7 col-md-7 col-lg-7 pull-left col-xs-offset-1 col-md-offset-1 col-sm-offset-1 col-lg-offset-1 panel panel-info post-{{$post->id}}">
+		<!-- Blogpost blok -->
+
+		<div class="col-xs-8 col-sm-8 col-md-8 col-lg-8 pull-left panel panel-info post-{{$post->id}}">
 			<h1>{{ $post->title }}  </h1>
-			<p class="postinfo">Geplaatst op {{ $post->getUpdatedAt() }} door <a href="{{{ URL::to('users/'.$user->username) }}}" target="_blank">{{ $author }}</a>
+			<p class="postinfo">Geplaatst op {{ $post->getUpdatedAt() }} door <a href="{{{ URL::to('user/'.$user->username) }}}" target="_blank">{{ $author }}</a>
 			</p>
 			<p>{{ $post->image }}</p>
 			<p>{{ $post->body }}</p>
 			<hr>
-			<div class="blogbottom">
-				<p class="pull-left">Tags: 
-			@foreach($post->tags()->get() as $tag)
-				<a href="{{ URL::to('tags/'.$tag->name) }}" title="{{$tag->name}}" alt="{{$tag->name}}">{{ $tag->name }}</a>,  
-			@endforeach<br />
-				Posted in 
+			<div class="blogbottom col-xs-12 col-sm-12 col-md-12 col-lg-12">
+				<div class="pull-left col-xs-7 col-sm-7 col-md-7 col-lg-7">
+					Tags: 
+					@foreach($post->tags as $tag)
+						<a href="{{ URL::to('tags/'.$tag->name) }}" title="{{$tag->name}}" alt="{{$tag->name}}">{{ $tag->name }}</a>,  
+					@endforeach<br />
+					Posted in 
 					@foreach($post->categories as $categorie)
 						<a href='categories/{{$categorie->name}}'>{{ $categorie->name }}</a> 
 					@endforeach
-				</p>
-				<p class="pull-right">
-					{{ Form::open(array('method'=>'DELETE','class'=>'pull-right','action'=>array('admin.blog.destroy', $post->slug))) }}
-					{{ Form::submit('Delete this post', array('class'=>'btn btn-danger btn-xs'))}}
-					{{ Form::close() }}
-				</p>
-				<p class="pull-right">
-					<a href="{{{ URL::route('admin.blog.edit', [$post->slug]) }}}" class="btn btn-warning btn-sm"><span class="glyphicon glyphicon-cog"></span> Edit this post</a>
-				</p>
-				<p class="pull-right">
-					<a href="{{{ URL::route('admin.tag.create') }}}" class="btn btn-primary btn-sm"><span class="glyphicon glyphicon-tag"></span> Add tag</a>
-				</p>
-				<p class="pull-right">
-					<a href="{{{ URL::route('admin.categorie.create') }}}" class="btn btn-info btn-sm"><span class="glyphicon glyphicon-tag"></span> Add Categorie</a>
-				</p>
+				</div>
+				@if($loggedUser->roles->first()->name === 'Admin')
+					<div class="pull-right col-xs-5 col-sm-5 col-md-5 col-lg-5">
+						{{ Form::open(array('method'=>'DELETE','class'=>'pull-right','action'=>array('admin.blog.destroy', $post->slug))) }}
+						{{ Form::submit('Delete this post', array('class'=>'btn btn-danger btn-xs'))}}
+						{{ Form::close() }}
+					
+				 
+						<a href="{{{ URL::route('admin.blog.edit', [$post->slug]) }}}" class="btn btn-warning btn-sm pull-right" style="margin-right: 10px;"><span class="glyphicon glyphicon-cog"></span> Edit this post</a>
+					</div>
+				@endif
 			</div>
+		</div><!-- /Blogpost blok -->
 
-			
-		</div>
+		<!-- Auteur blok -->
 
-		<div class="col-xs-7 col-sm-7 col-md-7 col-lg-7 pull-left col-xs-offset-1 col-md-offset-1 col-sm-offset-1 col-lg-offset-1 panel panel-info">
+		<div class="col-xs-8 col-sm-8 col-md-8 col-lg-8 panel panel-info">
 			<div class="col-xs-5 col-sm-5 col-md-5 col-lg-5 pull-left">
-				<img class="profile-pic" src="{{{ URL::to($user->image) }}}"></img>
+				<figure class="profile-pic" style="background-image: url('{{{ URL::to($user->image) }}}');"></figure>
 				<h3>{{$author}}</h3>
 			</div>
 			<div class="col-xs-7 col-sm-7 col-md-7 col-lg-7 pull-right">
 				<p>{{$user->description}}</p>
 			</div>
-		</div>
+		</div> <!-- /Auteur blok -->
+
+		<!-- Gerelateerde posts blok -->
+		@include('layouts.front.menus.blog.related')
+		<!-- /Gerelateerde posts blok -->
+
+		<!-- Comments blok -->
 
 		@foreach($post->comments as $comment)
-		<div class="col-xs-7 col-sm-7 col-md-7 col-lg-7 pull-left col-xs-offset-1 col-md-offset-1 col-sm-offset-1 col-lg-offset-1 panel panel-warning">
+		<div class="col-xs-7 col-sm-7 col-md-7 col-lg-7 pull-left panel panel-warning comment-{{$comment->id}}">
+			@if($loggedUser->roles->first()->name === 'Admin')
+			
+			<a href="{{ URL::route('comment.show',$comment->id) }}">Kijk</a>
+			
+			@endif
 			<div class="col-xs-5 col-sm-5 col-md-5 col-lg-5 pull-left">
 				<a href="{{URL::to($comment->website)}}" target="_blank"><h3>{{$comment->naam}}</h3></a>
 				<p class="pull-right">Geplaatst op: {{ $comment->created_at }}</p>
@@ -78,11 +80,16 @@
 		</div>
 		@endforeach
 
-		
-</div>
+		<!-- /Comments blok -->
 
-@stop
 
-@section('comments')
+<!-- Nieuwe comment blok -->
+<div class="nieuw-comment col-xs-8 col-sm-8 col-md-8 col-lg-8">
+	<h3>Plaats een reactie:</h3>
 	@include('comments.form')
+</div><!-- /Nieuwe comment blok -->
+	
+</div><!-- /Blogrol.container -->
+
 @stop
+
